+++ volgrids/smiffer/_smifs/hydrophobic.py
@@ -15,7 +15,5 @@ class SmifHydrophobic(SmifHydro):
             if mul_factor < 0: continue
             kernel.stamp(grid, particle.position, multiply_by =  mul_factor)
 
-        grid.dirty = True
-
 
 # //////////////////////////////////////////////////////////////////////////////