+++ volgrids/smutils/app_smutils.py
@@ -16,6 +16,7 @@ class AppSMUtils(vg.AppSubcommand):
         if operation == "res_nobp" : return self._run_res_nobp()
         if operation == "res_nostk": return self._run_res_nostk()
         if operation == "chemgen"  : return self._run_chemgen()
+        if operation == "sphere"   : return self._run_sphere()
         if operation == "occupancy": return self._run_occupancy()
         if operation == "pwoverlap": return self._run_pwoverlap()
         if operation == "log_apbs" : return self._run_log_apbs()
@@ -36,13 +37,18 @@ class AppSMUtils(vg.AppSubcommand):
 
     # --------------------------------------------------------------------------
     def _run_chemgen(self) -> None:
-        path_in = self.main.get_arg_path("path_in", assertion = fy.PathAssertion.FILE_IN)
+        path_in  = self.main.get_arg_path("path_in", assertion = fy.PathAssertion.FILE_IN)
         path_out = self.main.get_arg_path("path_out",
             assertion = fy.PathAssertion.FILE_OUT, default = path_in.with_suffix(".chem")
         )
         path_out.write_text(
             su.ChemTableLigand(path_in).gen_chemtable()
         )
+
+
+    # --------------------------------------------------------------------------
+    def _run_sphere(self) -> None:
+        su.AppSpheres(self.main).run()
 
 
     # --------------------------------------------------------------------------